[BDAP] Move leveldb code to seperate code file
